Use .js instead of .cjs
- add new router method to handle all routes with a prefix /api
