+++ db/connection.js
@@ -1,11 +1,9 @@
-const sqlite3 = require("sqlite3").verbose();
-const md5 = require("md5");
-const { OPEN_READWRITE } = require("sqlite3");
-const fs = require("fs");
+import sqlite3 from "sqlite3";
+import { access, constants } from "fs";
 
 const dbSource = "./db/wp-ctp.db";
 
-fs.access(dbSource, fs.constants.F_OK, (err) => {
+access(dbSource, constants.F_OK, (err) => {
   if (err) {
     console.error(`Database file does not exist at ${dbSource}`);
     return;
@@ -13,7 +11,7 @@ fs.access(dbSource, fs.constants.F_OK, (err) => {
 });
 
 let db = new sqlite3.Database(dbSource, sqlite3.OPEN_READWRITE, (err) => {
-  if (err || dbSource=="") {
+  if (err || dbSource == "") {
     // Cannot open database
     console.error(err.message);
     throw err;
@@ -22,4 +20,4 @@ let db = new sqlite3.Database(dbSource, sqlite3.OPEN_READWRITE, (err) => {
   }
 });
 
-module.exports = db;
+export default db;
