--- src/index.cjs
@@ -1,5 +1,0 @@
-require("dotenv").config();
-const Server = require("./server.cjs");
-const app = new Server();
-
-app.listen();
