+++ src/index.js
@@ -1,0 +1,6 @@
+import { config } from "dotenv";
+config();
+import Server from "./server.js";
+const app = new Server();
+
+app.listen();
